Set "scalaVersion in ThisBuild".
This ensures that if a parent POM sets a scala version, but the children POMs
do not inherit that scala.version property for whatever nefarious reason, SBT
still uses the parent scala version for all subprojects.

Otherwise SBT will pick a Scala version out of its ass for the subprojects
because SBT loves it when hilarity ensues.